Move MemoryRepr to VM
